code/requirements.txt
google-genai==1.29.0
mistralai==1.9.3
numpy==2.3.1
pandas==2.3.1